Move testActivityMenuRequestDesktopItem to BrowsingTest. Add TestRail link.
