+++ focus-ios/focus-ios-tests/XCUITest/BrowsingTest.swift
@@ -75,4 +75,38 @@ class BrowsingTest: BaseTestCase {
         app.buttons["BrowserToolset.stopReloadButton"].tap()
         waitForValueContains(app.textFields["URLBar.urlText"], value: "mozilla")
     }
+
+    // Smoketest
+    // https://testrail.stage.mozaws.net/index.php?/cases/view/2587661
+    func testActivityMenuRequestDesktopItem() {
+        let urlBarTextField = app.textFields["URLBar.urlText"]
+
+        // Wait for existence rather than hittable because the textfield is technically disabled
+        loadWebPage("facebook.com")
+
+        waitForWebPageLoad()
+        waitForExistence(app.buttons["HomeView.settingsButton"])
+        app.buttons["HomeView.settingsButton"].tap()
+
+        if iPad() {
+            waitForExistence(app.collectionViews.buttons["Request Mobile Site"])
+            app.collectionViews.buttons["Request Mobile Site"].tap()
+        } else {
+            waitForExistence(app.collectionViews.buttons["Request Desktop Site"])
+            app.collectionViews.buttons["Request Desktop Site"].tap()
+        }
+
+        waitForWebPageLoad()
+
+        guard let text = urlBarTextField.value as? String else {
+            XCTFail()
+            return
+        }
+
+        if text.contains("m.facebook") {
+            if !iPad() {
+                XCTFail()
+            }
+        }
+    }
 }
